chore: install typescript
