+++ package.json
@@ -9,6 +9,10 @@
 		"@testing-library/jest-dom": "^5.16.5",
 		"@testing-library/react": "^13.4.0",
 		"@testing-library/user-event": "^14.4.3",
+		"@types/jest": "^29.5.3",
+		"@types/node": "^20.5.1",
+		"@types/react": "^18.2.20",
+		"@types/react-dom": "^18.2.7",
 		"firebase": "^9.15.0",
 		"formik": "^2.2.9",
 		"framer-motion": "^7.10.2",
@@ -21,6 +25,7 @@
 		"react-router-dom": "^6.5.0",
 		"react-scripts": "5.0.1",
 		"redux": "^4.2.0",
+		"typescript": "^5.1.6",
 		"web-vitals": "^3.1.0",
 		"yup": "^0.32.11"
 	},
